copy_dataset/requirements.txt
google-cloud-bigquery~=3.35
